server/requirements.txt
flask>=3.0
flask-cors>=4.0
Pillow>=10.0
numpy>=1.24
open-clip-torch>=2.24
torch>=2.1
transformers>=4.36
facenet-pytorch>=2.5
insightface>=0.7
scikit-learn>=1.3
